Refactored ping functionality
